config.py
import os

from dotenv import load_dotenv

load_dotenv()


class Config:
    DEBUG = False
    TESTING = False
    ALPHA_VANTAGE_API_KEY = os.getenv("ALPHA_VANTAGE_API_KEY")


class DevConfig(Config):
    DEBUG = True


class TestConfig(Config):
    TESTING = True
