requirements.txt
Flask==1.1.1
Flask-RESTful==0.3.7
pandas-datareader==0.8.1
python-dotenv==0.10.3
